Simplify datetime mutator
